+++ include/Ark/FFI.hpp
@@ -3,6 +3,8 @@
 
 #include <vector>
 #include <string>
+#include <iostream>
+#include <sstream>
 
 namespace Ark
 {